Machine Learning with Cloud Computing/Practical 5/requirements.txt
pandas==2.2.0
python-dotenv==1.0.1
scipy==1.13.0